+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.3.0 (unreleased)
 
+- Raise `ArgumentError` for too many arguments with enumerable
 - Dropped support for Active Record < 5.2 and Ruby < 2.6
 
 ## 0.2.4 (2020-09-07)